Refactored out spreadsheet handling code
